+++ src/main/java/com/shuidi/zhanggui/service/bl/impl/SpecificationsServiceImpl.java
@@ -28,7 +28,7 @@ public class SpecificationsServiceImpl implements SpecificationsService {
   }
 
   @Override
-  public int insertSpecifications(Specifications t) {
+  public Long insertSpecifications(Specifications t) {
     return specificationsDao.insertSpecifications(t);
   }
